+++ dev/shiny-input-list/ui.R
@@ -1,0 +1,25 @@
+
+
+# UI ----------------------------------------------------------------------
+
+
+ui <- fluidPage(
+
+  # classic app
+  headerPanel('Iris k-means clustering'),
+  sidebarPanel(
+    selectInput('xcol', 'X Variable', names(iris)),
+    selectInput('ycol', 'Y Variable', names(iris),
+                selected=names(iris)[[2]]),
+    numericInput('clusters', 'Cluster count', 3,
+                 min = 1, max = 9)
+  ),
+  mainPanel(
+    plotOutput('plot1'),
+    verbatimTextOutput("res_auth")
+  )
+
+)
+
+secure_app(ui, enable_admin = TRUE, theme = shinythemes::shinytheme("sandstone"))
+